requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
python-dotenv==1.0.0
chromadb==0.4.22
sentence-transformers==2.2.2
gitpython==3.1.40
openai>=1.12.0
pydantic==2.6.1
numpy==1.26.4